requirements.txt
openai>=1.12.0
tiktoken>=0.5.0
chromadb>=0.4.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
streamlit>=1.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
tqdm>=4.66.0
loguru>=0.7.0